// frontend/src/models/response/IUser.ts
export type IUser = {
  id: number;
  data: {
    firstname: string;
    secondname: string;
    lastname: string;
    nickname: string;
    team: string;
    team_role: string;
    email: string;
    organization: string;
    avatar_path: string;
    bio: string;
    ctf_compete: string;
    task_complete: string;
    contact: string;
  };
};